HttpRequest: fix gcc2 build after f4fce2fd2e81039b247b2fee90a69a98490742ca

+++ src/kits/network/libnetapi/HttpRequest.cpp
@@ -750,7 +750,7 @@ BHttpRequest::_MakeRequest()
 							bytesReceived - bytesRead, bytesRead);
 					}
 					fListener->DownloadProgress(this, bytesReceived,
-						std::max((ssize_t)0, bytesTotal));
+						std::max((off_t)0, bytesTotal));
 				}
 
 				if (bytesTotal >= 0 && bytesReceived >= bytesTotal)
